refactor(account): split account_creation_manager into class + api (audit T4.6)
account_creation_manager.py was a 2,260-LOC file gluing two
responsibilities together. Split along the existing clean seam:

- account_creation_manager.py keeps the AccountCreationManager pipeline
  class (~1,024 LOC). 8 imports that only the moved functions used are
  removed.
- account_creation_api.py (new) holds the 9 @frappe.whitelist() queue /
  bulk / retry endpoints (~1,180 LOC) that drive the class.

Verbatim move — no logic changed. Backward compatibility:

- The deprecated shim verenigingen/utils/account_creation_manager.py now
  re-exports from both modules, so every old-path import (and the
  frappe.enqueue dotted-string paths) keeps working.
- volunteer_activation_service.py imported upgrade_member_to_volunteer_user
  from the services path directly — repointed to account_creation_api.
- frappe.enqueue dotted-string targets for process_account_creation_request
  / process_bulk_account_creation_batch repointed to the new module
  (same function, avoids a deprecation warning per enqueue).
- whitelist_files.txt: added account_creation_api.py.
