+++ verenigingen/services/volunteer/volunteer_activation_service.py
@@ -79,7 +79,7 @@ def activate_volunteer_record(member):
             # Upgrade user account from Website User to System User for volunteer access
             if member.user:
                 try:
-                    from verenigingen.services.member.account.account_creation_manager import (
+                    from verenigingen.services.member.account.account_creation_api import (
                         upgrade_member_to_volunteer_user,
                     )
 
@@ -110,7 +110,7 @@ def activate_volunteer_record(member):
                 # Upgrade user account from Website User to System User for volunteer access
                 if member.user:
                     try:
-                        from verenigingen.services.member.account.account_creation_manager import (
+                        from verenigingen.services.member.account.account_creation_api import (
                             upgrade_member_to_volunteer_user,
                         )
