+++ verenigingen/utils/account_creation_manager.py
@@ -2,8 +2,10 @@
 import warnings
 
 warnings.warn(
-    "Import from verenigingen.services.member.account.account_creation_manager instead",
+    "Import from verenigingen.services.member.account.account_creation_manager "
+    "(AccountCreationManager class) or account_creation_api (whitelisted endpoints) instead",
     DeprecationWarning,
     stacklevel=2,
 )
+from verenigingen.services.member.account.account_creation_api import *  # noqa: E402,F401,F403
 from verenigingen.services.member.account.account_creation_manager import *  # noqa: E402,F401,F403
